structure decomposed
