+++ src/Components/Movies/Features/Movies/Movies.jsx
@@ -1,9 +1,10 @@
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchMovies, setCategory } from "../../Redux/moviesSlice";
+import { fetchMovies, setCategory } from "../../../../Redux/Features/Movies/moviesSlice";
 import { useNavigate } from "react-router-dom";
 import { Grid, CardMedia, CardContent, CircularProgress, Box } from "@mui/material";
-import { MoviesContainer, MovieCard, MovieTitle, MovieOverview, CategoryButton } from "./MoviesStyles";
+import { MoviesContainer, MovieCard, MovieTitle, MovieOverview, CategoryButton, TvTitle } from "./MoviesStyles";
+import TvShows from "../TvShows/TvShows";
 
 const Movies = () => {
     const dispatch = useDispatch();
@@ -54,6 +55,8 @@ const Movies = () => {
                     </Grid>
                 ))}
             </Box>
+
+            <TvShows/>
         </MoviesContainer>
     );
 };
