+++ src/Redux/redux-store.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
-import moviesReducer from "./moviesSlice";
-import movieItemReducer from "./moviesItemSlice";
-import actorsReducer from "./actorsSlice";
+import moviesReducer from "./Features/Movies/moviesSlice";
+import movieItemReducer from "./Features/Movies/moviesItemSlice";
+import actorsReducer from "./Features/Movies/actorsSlice";
 
 export const store = configureStore({
     reducer: {
